Refactoring Updated code with OpenAPI generator Moved FormsService to directory
