+++ src/app/services/fn/requests/get-incoming-requests.ts
@@ -9,7 +9,6 @@ import { RequestBuilder } from '../../request-builder';
 import { PageRequest } from '../../models/page-request';
 
 export interface GetIncomingRequests$Params {
-  location: string;
   page: number;
   size: number;
 }
@@ -17,7 +16,6 @@ export interface GetIncomingRequests$Params {
 export function getIncomingRequests(http: HttpClient, rootUrl: string, params: GetIncomingRequests$Params, context?: HttpContext): Observable<StrictHttpResponse<PageRequest>> {
   const rb = new RequestBuilder(rootUrl, getIncomingRequests.PATH, 'get');
   if (params) {
-    rb.query('location', params.location, {});
     rb.query('page', params.page, {});
     rb.query('size', params.size, {});
   }
