+++ src/app/services/forms/forms.service.spec.ts
@@ -1,6 +1,6 @@
-import { TestBed } from '@angular/core/testing';
+import {TestBed} from '@angular/core/testing';
 
-import { FormsService } from './forms.service';
+import {FormsService} from './forms.service';
 
 describe('FormsService', () => {
   let service: FormsService;
